requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bert-score==0.3.11
bleach==5.0.1
blis==0.7.8
boto3==1.24.17
botocore==1.27.17
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
easydict==1.9
editdistpy==0.1.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.6.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
murmurhash==1.0.7
nbclient==0.6.5
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nlpaug==1.1.10
nltk==3.7
notebook==6.4.12
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plyfile==0.7.4
pointnet2==3.0.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.4
pycparser==2.21
pydantic==1.8.2
pyDeprecate==0.3.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.1
PyWavefront==1.3.3
PyYAML==6.0
pyzmq==23.2.0
Quaternion==3.5.2.post4
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.3.0
symspellpy==6.7.6
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.15.0
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
trimesh==2.35.39
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
yarl==1.7.2
zipp==3.8.0
